requirements.txt
python-telegram-bot==13.15
openai==1.23.2
python-dotenv==1.0.1
schedule==1.2.1